read/write issue resolved for > 3.5 .net framework

+++ SharpBot/MacroBot/Protocol/Planet.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public int OrbitDistance { get; set; }
-        public int OrbitRotation { get; set; }
+        public double OrbitRotation { get; set; }
         public int OwnedBy { get; set; }
     }
 }
